first batch in proces to update to Elgg coding standards

+++ procedures/question/delete.php
@@ -3,12 +3,10 @@
 //annotationId
 $returnData = array();
 
-$question_guid = get_input("guid");
+$question_guid = (int) get_input("guid");
 
-if(!empty($question_guid) && $question = get_entity($question_guid))
-{
-	if(!empty($question) && ($question instanceof EventRegistrationQuestion))
-	{
+if (!empty($question_guid) && $question = get_entity($question_guid)) {
+	if (!empty($question) && ($question instanceof EventRegistrationQuestion)) {
 		$question->delete();
 	}
 	$returnData['valid'] = 1;
