+++ procedures/slot/delete.php
@@ -4,19 +4,15 @@ $returnData = array();
 
 $returnData['valid'] = 0;
 
-$guid = get_input("guid");
+$guid = (int) get_input("guid");
 
-if(!empty($guid) && $eventSlot = get_entity($guid))
-{
-	if($eventSlot->getSubtype() == EventSlot::SUBTYPE)
-	{
-		if($eventSlot->delete())
-		{
+if (!empty($guid) && $eventSlot = get_entity($guid)) {
+	if ($eventSlot->getSubtype() == EventSlot::SUBTYPE) {
+		if ($eventSlot->delete()) {
 			$returnData['valid'] = 1;
 		} 
 	}
 }
 
 echo json_encode($returnData);
-
 exit;
